+++ src/engine/w_file.h
@@ -57,7 +57,7 @@ struct _wad_file_s {
 
 	// Length of the file, in bytes.
 
-	uint64_t length;
+	intptr_t length;
 };
 
 // Open the specified file. Returns a pointer to a new wad_file_t
